tx: move features/util -> capabilties/generic
