+++ packages/tx/src/2930/tx.ts
@@ -9,8 +9,8 @@ import {
 
 import * as EIP2718 from '../capabilities/eip2718.js'
 import * as EIP2930 from '../capabilities/eip2930.js'
+import { getBaseJSON, sharedConstructor, valueBoundaryCheck } from '../capabilities/generic.js'
 import * as Legacy from '../capabilities/legacy.js'
-import { getBaseJSON, sharedConstructor, valueBoundaryCheck } from '../features/util.js'
 import { TransactionType } from '../types.js'
 import { AccessLists } from '../util.js'
